# docs/swagger.yaml
definitions:
...
  request.ProductItemReq:
    properties:
      discount_price:
        type: integer
      price:
        type: integer
      product_brand:
        type: string
      product_id:
        type: integer
      qty_in_stock:
        type: integer
    required:
    - product_brand
    - product_id
    - qty_in_stock
    type: object
  request.ProductReq:
    properties:
      category_id:
        type: integer
      description:
        maxLength: 1000
        minLength: 10
        type: string
      product_name:
        maxLength: 50
        minLength: 3
        type: string
    required:
    - category_id
    - description
    - product_name
    type: object
...
  request.UpdateProductReq:
    properties:
      brand_id:
        type: integer
      description:
        type: string
      id:
        type: integer
      product_name:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      errors: {}
      message:
        type: string
      status_code:
        type: integer
    type: object
info:
  contact:
    name: API Support
  description: Glamgrove is an E-commerce platform for purchasing high quality beauty
    products
  title: Go + Gin E-Commerce API Glamgrove
  version: 1.0.0
paths:
  /admin/brands/add:
    post:
      consumes:
      - application/json
      description: Adds a new product to the database.
      parameters:
      - description: Product details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: Success message with product details
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to add product
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add a new product
      tags:
      - Products
  /admin/brands/get:
    get:
      consumes:
      - application/json
      description: Retrieves all categories from the database.
      produces:
      - application/json
      responses:
        "200":
          description: Success message with all categories
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get categories
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all categories
      tags:
      - Categories
...
  /admin/dashboard/salesdata:
    get:
      consumes:
      - application/json
      description: Retrieves sales statistics within the specified date range.
      parameters:
      - description: Start date (YYYY-MM-DD)
        in: query
        name: startDate
        required: true
        type: string
      - description: End date (YYYY-MM-DD)
        in: query
        name: endDate
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Sales data fetched successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Can't calculate details of sales
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get sales statistics
      tags:
      - Products
...
  /admin/products/additemimage:
    post:
      consumes:
      - multipart/form-data
      description: Adds image(s) to a specific product item.
      parameters:
      - description: Product Item ID
        in: formData
        name: product_item_id
        required: true
        type: integer
      - description: Image file(s)
        in: formData
        name: image
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: Success message with added images
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Error message with details
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Add image to a product item
      tags:
      - Products
  /admin/products/delete:
    delete:
      consumes:
      - application/json
      description: Deletes the product with the specified ID.
      parameters:
      - description: Product ID to delete
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product deleted successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Missing or invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to delete product
          schema:
            $ref: '#/definitions/response.Response'
      summary: Delete a product
      tags:
      - Products
...
  /admin/products/list:
    get:
      consumes:
      - application/json
      description: Retrieves a list of products with pagination support.
      parameters:
      - description: Number of products to retrieve per page
        in: query
        name: count
        type: integer
      - description: Page number
        in: query
        name: page_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No products to show
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get all products
          schema:
            $ref: '#/definitions/response.Response'
      summary: List products
      tags:
      - Products
  /admin/products/product-item:
    post:
      consumes:
      - application/json
      description: Adds a new product item.
      parameters:
      - description: Product item details
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.ProductItemReq'
      produces:
      - application/json
      responses:
        "200":
          description: Product item added successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Missing or invalid input
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to add product item
          schema:
            $ref: '#/definitions/response.Response'
      summary: Add a product item
      tags:
      - Products
  /admin/products/product-item/:product_id:
    get:
      consumes:
      - application/json
      description: Retrieves product items based on the provided product ID.
      parameters:
      - description: Product ID
        in: query
        name: product_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Product items fetched successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid param input" or "No product items for this product
            ID
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get product item for given product ID
          schema:
            $ref: '#/definitions/response.Response'
      summary: Get product items by product ID
      tags:
      - Products
  /admin/products/update:
    put:
      consumes:
      - application/json
      description: Updates an existing product with the provided details.
      parameters:
      - description: Product ID to update
        in: path
        name: id
        required: true
        type: integer
      - description: Product details to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateProductReq'
      produces:
      - application/json
      responses:
        "200":
          description: Product updated successfully
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Failed to update product
          schema:
            $ref: '#/definitions/response.Response'
      summary: Update a product
      tags:
      - Products
...
  /products/:
    get:
      consumes:
      - application/json
      description: Retrieves a list of products with pagination support.
      parameters:
      - description: Number of products to retrieve per page
        in: query
        name: count
        type: integer
      - description: Page number
        in: query
        name: page_number
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: No products to show
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Invalid inputs
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get all products
          schema:
            $ref: '#/definitions/response.Response'
      summary: List products
      tags:
      - Products
  /products/brands:
    get:
      consumes:
      - application/json
      description: Retrieves all categories from the database.
      produces:
      - application/json
      responses:
        "200":
          description: Success message with all categories
          schema:
            $ref: '#/definitions/response.Response'
        "500":
          description: Failed to get categories
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: Get all categories
      tags:
      - Categories
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"